requirements.txt
attrs==21.2.0
aws-cdk.assets==1.121.0
aws-cdk.aws-apigateway==1.121.0
aws-cdk.aws-applicationautoscaling==1.121.0
aws-cdk.aws-autoscaling==1.121.0
aws-cdk.aws-autoscaling-common==1.121.0
aws-cdk.aws-autoscaling-hooktargets==1.121.0
aws-cdk.aws-certificatemanager==1.121.0
aws-cdk.aws-cloudformation==1.121.0
aws-cdk.aws-cloudfront==1.121.0
aws-cdk.aws-cloudwatch==1.121.0
aws-cdk.aws-codebuild==1.121.0
aws-cdk.aws-codecommit==1.121.0
aws-cdk.aws-codeguruprofiler==1.121.0
aws-cdk.aws-codepipeline==1.121.0
aws-cdk.aws-codestarnotifications==1.121.0
aws-cdk.aws-cognito==1.121.0
aws-cdk.aws-ec2==1.121.0
aws-cdk.aws-ecr==1.121.0
aws-cdk.aws-ecr-assets==1.121.0
aws-cdk.aws-ecs==1.121.0
aws-cdk.aws-ecs-patterns==1.121.0
aws-cdk.aws-efs==1.121.0
aws-cdk.aws-elasticloadbalancing==1.121.0
aws-cdk.aws-elasticloadbalancingv2==1.121.0
aws-cdk.aws-events==1.121.0
aws-cdk.aws-events-targets==1.121.0
aws-cdk.aws-globalaccelerator==1.121.0
aws-cdk.aws-iam==1.121.0
aws-cdk.aws-kinesis==1.121.0
aws-cdk.aws-kinesisfirehose==1.121.0
aws-cdk.aws-kms==1.121.0
aws-cdk.aws-lambda==1.121.0
aws-cdk.aws-logs==1.121.0
aws-cdk.aws-route53==1.121.0
aws-cdk.aws-route53-targets==1.121.0
aws-cdk.aws-s3==1.121.0
aws-cdk.aws-s3-assets==1.121.0
aws-cdk.aws-sam==1.121.0
aws-cdk.aws-secretsmanager==1.121.0
aws-cdk.aws-servicediscovery==1.121.0
aws-cdk.aws-signer==1.121.0
aws-cdk.aws-sns==1.121.0
aws-cdk.aws-sns-subscriptions==1.121.0
aws-cdk.aws-sqs==1.121.0
aws-cdk.aws-ssm==1.121.0
aws-cdk.aws-stepfunctions==1.121.0
aws-cdk.cloud-assembly-schema==1.121.0
aws-cdk.core==1.121.0
aws-cdk.custom-resources==1.121.0
aws-cdk.cx-api==1.121.0
aws-cdk.region-info==1.121.0
cattrs==1.8.0
-e git+https://github.com/qzhuyan/cdk_mon.git@4c28ae672cd2b73454befa8be8c651858fd8742f#egg=cdk_mon
constructs==3.3.141
jsii==1.34.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing-extensions==3.10.0.2
